Start of RateQueue implementation with Pekko
What:
This code creates a QueueOfQueues actor and for each enqueue stores
the json payload into a PhoneQueue actor. QueueOfQueues keeps a
priority queue and a hash map of active PhoneQueue references. The
first element in the queue gets a Dequeue request in which it either:

1. Fails
2. Returns Next Json Payload
3. Returns Next Json Payload and an Empty flag

On Next Json Payload reponse, a new item is inserting to the
QueueOfQueues that represents the next second because there are more
items in the underlying actor

On Empty Flag or Fails, the Actor is assumed to be Behaviors.stop and
therefore, is cleaned up from Hash Map