Transactor Behavior implementation
What:
The transactor wraps the QueueOfQueues and just calls enqueue on a new
event. On a dequeue request, it dequeues from the QueueOfQueues and
opens a transaction with a timeout of 10 seconds. In the event that
the  `CompleteTransaction` message isn't received in 10 seconds, the
transactions is closed and the object is reenqueued -- sadly at the
current timestamp which is not the best.

In the event a `CompleteTransaction` message is received for the open
transaction, the pending failure message is canceled and the
transaction is removed from the `activeTransactions` state.