Move PhoneQueue to pekko persistence library
Why:
This is step 1 of moving to fully sharded and durable queues. The
persistence library allows for actors to save their state in a
database as a journal and as a snapshot.

How:
Rewrite the PhoneQueue behavior entirely but keep the tests. The test
guide the implementation and make sure that we setup the right
refactoring to make this possible