Move Transactor to Persistence system
Why:
This is in prep for doing a full cluster and persistence deployment of
the system. Goal is to make it so everything is recoverable from
journal and snapshot

How:
Store the state as a map of transactions by id. The state can change
by removing a small set of transactions or adding a single
transaction. After manipulating the state, the system then sends
messages to the respective actors / replies