Test for the queue delay
Why:
After reviewing the code, it became clear that this always returned a queue item
even if that queue item was too young. In the test for the same message twice,
the there was no waiting for the required duration

How:
First wrote the test to fish for the message and assert that there are several
Empty responses before the next event.

Then fixed the logic to handle when the queue has an item but that time is too
young to be used.